+++ src/lib/components/common/Button/index.ts
@@ -1,0 +1,3 @@
+import Button from './Button.svelte';
+
+export default Button;
